# iCanteen-api接口.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "iCanteen-api接口",
    "description": "校园“智汇”食堂后端接口文档（返回统一使用 Result 包装）",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/dish/{id}": {
      "get": {
        "summary": "免登录查询菜品详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "菜品"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "菜品ID",
            "required": true,
            "example": 6,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=Dish)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除菜式（管理员）",
        "deprecated": false,
        "description": "",
        "tags": [
          "菜品"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "菜品ID",
            "required": true,
            "example": 14,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/dish/recommend/by-canteen": {
      "get": {
        "summary": "免登录获取某食堂推荐菜式",
        "deprecated": false,
        "description": "输入食堂ID，返回四道菜详情：1道该食堂最高分中随机菜式 + 3道管理员手动推荐菜式",
        "tags": [
          "菜品"
        ],
        "parameters": [
          {
            "name": "canteenId",
            "in": "query",
            "description": "食堂ID",
            "required": true,
            "example": 2,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=Dish[])",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/dish/recommend/manual": {
      "post": {
        "summary": "手动推荐三道菜式（管理员）",
        "deprecated": false,
        "description": "输入食堂id和属于该食堂的3个菜式id，设置成功后返回提示",
        "tags": [
          "菜品"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishManualRecommendDTO"
              },
              "example": {
                "canteenId": 2,
                "dishIds": [
                  7,
                  9,
                  11
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result(data=设置成功)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/dish": {
      "post": {
        "summary": "增加菜式（管理员）",
        "deprecated": false,
        "description": "必须有窗口id，不需要菜式id（自增生成）",
        "tags": [
          "菜品"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishCreateDTO"
              },
              "example": {
                "name": "汤粉",
                "description": "暖呼呼一碗",
                "windowId": 6,
                "image": "/imgs/dish/deliciousnoodle.jpg",
                "isToday": 1,
                "price": 18.66,
                "quantity": 99,
                "status": 1,
                "tags": "noodle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result(data=id)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "修改菜式（管理员）",
        "deprecated": false,
        "description": "必须有菜式id",
        "tags": [
          "菜品"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishUpdateDTO"
              },
              "example": {
                "id": 14,
                "name": "老干妈拌饭",
                "description": "妈妈的味道~",
                "windowId": 7,
                "image": "/imgs/dish/yeyerice.jpg",
                "isToday": 0,
                "price": 12.88,
                "quantity": 100,
                "status": 1,
                "tags": "great"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/dish-review": {
      "post": {
        "summary": "已登录且已下过单用户提交菜品评价（打分+评语）",
        "deprecated": false,
        "description": "用户提交菜品评价成功后奖励3积分。",
        "tags": [
          "菜品评价"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "9653930b1b53492a80214bbca92c77a3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishReviewSubmitDTO"
              },
              "example": {
                "dishId": 1,
                "rating": 5,
                "content": "口味不错，分量足"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/dish-review/by-dish": {
      "get": {
        "summary": "用户分页查看某道菜品评价",
        "deprecated": false,
        "description": "",
        "tags": [
          "菜品评价"
        ],
        "parameters": [
          {
            "name": "dishId",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "9653930b1b53492a80214bbca92c77a3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=DishReview[],total=总数)",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/dish-review/my": {
      "get": {
        "summary": "用户分页查看自己发表过的菜品评价",
        "deprecated": false,
        "description": "",
        "tags": [
          "菜品评价"
        ],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "9653930b1b53492a80214bbca92c77a3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=DishReview[],total=总数)",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/dish-review/{id}": {
      "delete": {
        "summary": "用户删除发表过的菜品评价",
        "deprecated": false,
        "description": "",
        "tags": [
          "菜品评价"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "评价ID",
            "required": true,
            "example": 5,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "9653930b1b53492a80214bbca92c77a3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "Result": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errorMsg": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "null"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        }
      },
...
      "DishManualRecommendDTO": {
        "type": "object",
        "description": "管理员手动推荐菜式入参，必须传食堂id和3个菜品id",
        "required": [
          "canteenId",
          "dishIds"
        ],
        "properties": {
          "canteenId": {
            "type": "integer",
            "format": "int64"
          },
          "dishIds": {
            "type": "array",
            "minItems": 3,
            "maxItems": 3,
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "DishReviewSubmitDTO": {
        "type": "object",
        "description": "菜品评价入参（打分+评语）",
        "required": [
          "dishId",
          "rating",
          "content"
        ],
        "properties": {
          "dishId": {
            "type": "integer",
            "format": "int64"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "tags": {
            "type": "string"
          },
          "images": {
            "type": "string"
          }
        }
      },
...
      "DishCreateDTO": {
        "type": "object",
        "description": "管理员新增菜品参数（不需要id）",
        "required": [
          "windowId",
          "name",
          "price"
        ],
        "properties": {
          "windowId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0
          },
          "tags": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "isToday": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          }
        }
      },
      "DishUpdateDTO": {
        "type": "object",
        "description": "管理员修改菜品参数（必须携带id，其他字段按需修改）",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "windowId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0
          },
          "tags": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          },
          "isToday": {
            "type": "integer",
            "enum": [
              0,
              1
            ]
          }
        }
      }
    },
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}